fix: fix entire connection and disconnection process

+++ src/services/presence-manager.ts
@@ -4,6 +4,10 @@ import {IdleManager} from "./idle-manager";
 import {getEditorImage} from "../utils";
 import {SettingsManager} from "./settings-manager";
 import {EventEmitter} from "events";
+import { ConnectionStatus } from "../types/connection-status";
+
+type PresenceManagerEvents = ConnectionStatus | "disconnected";
+
 export default class PresenceManager extends EventEmitter {
   private updateInterval: NodeJS.Timeout | null = null;
 
@@ -16,18 +20,34 @@ export default class PresenceManager extends EventEmitter {
 
   private activityManager: ActivityManager;
 
+  private connectionStatus: ConnectionStatus = ConnectionStatus.USER_DISCONNECTED;
+
+    // Overriding emit to restrict event types
+    public emit(event: PresenceManagerEvents, ...args: any[]): boolean {
+      return super.emit(event, ...args);
+    }
+
   constructor(clientId: string) {
     super();
     this.client = new Client({
       clientId,
+      transport: {
+        type: "ipc",
+      }
     });
 
-    this.client.on("connected", () => {
-      this.emit("connected");
+    this.client.on("connected", async () => {
+      await this.changeConnectionStatus(ConnectionStatus.CONNECTED);
     });
 
-    this.client.on("disconnected", () => {
-      this.emit("disconnected");
+    this.client.on("disconnected", async () => {
+      console.log("Received disconnection event with status:", this.connectionStatus);
+
+      // If the connection status is connected, it means that the connection was unexpectedly disrupted. 
+      // Any other kind of disconnection would change the status first.
+      if(this.connectionStatus === ConnectionStatus.CONNECTED) {
+        await this.changeConnectionStatus(ConnectionStatus.LOST_UNEXPECTEDLY);
+      }
     });
 
     this.idleManager = new IdleManager(SettingsManager.instance.getIdleTimeout());
@@ -38,19 +58,37 @@ export default class PresenceManager extends EventEmitter {
     if (SettingsManager.instance.getDisconnectOnIdle()) {
       this.idleManager.onIdleChange((isIdle) => {
         if (isIdle) {
-          this.stopAndDisconnect();
+          if(this.connectionStatus === ConnectionStatus.CONNECTED) {
+            this.stopAndDisconnect(true);
+          }
         } else {
-          this.connectAndStartUpdating();
+          if(this.connectionStatus === ConnectionStatus.SYSTEM_DISCONNECTED) {
+            this.connectAndStartUpdating();
+          }
         }
       });
     }
 
-    this.on("disconnected", () => {
+    this.on(ConnectionStatus.LOST_UNEXPECTEDLY, () => {
+      console.log("RPC unexpectedly Disconnected");
       this.stopUpdating();
       if (SettingsManager.instance.getRetryConnection()) {
+        console.log("Starting retry connection");
         this.startRetryConnection();
       }
     });
+  }
+
+  private async changeConnectionStatus(status: ConnectionStatus, statusChangeCallback: () => Promise<void> = async () => {}) {
+    this.connectionStatus = status;
+    
+    await statusChangeCallback();
+    this.emit(status);
+
+    // Do not emit the "disconnected" event (used as a general event to sync UI) if system disconnected because we do not treat it as a proper disconnection but rather as a state
+    if(status !== ConnectionStatus.CONNECTED && status !== ConnectionStatus.SYSTEM_DISCONNECTED) {
+      this.emit("disconnected");
+    }
   }
 
   public async connectAndStartUpdating() {
@@ -59,30 +97,40 @@ export default class PresenceManager extends EventEmitter {
     }
 
     
-    await this.connectToClient().catch((error) => {
-      this.emit("disconnected");
+    await this.connectToClient().catch(async (error) => {
+      await this.changeConnectionStatus(ConnectionStatus.LOST_UNEXPECTEDLY);
       throw error;
     });
 
     this.startUpdating();
   }
 
-  public async stopAndDisconnect() {
+  public async stopAndDisconnect(isSystemDisconnected: boolean = false) {
     this.stopUpdating();
-    await this.disconnectFromClient();
+    await this.disconnectFromClient(isSystemDisconnected);
   }
 
   private async connectToClient() {
     await this.client.connect();
+    console.log("Connected to client");
   }
 
-  private async disconnectFromClient() {
-    await this.client.destroy();
+  private async disconnectFromClient(isSystemDisconnected: boolean = false) {
+    const newStatus = isSystemDisconnected ? ConnectionStatus.SYSTEM_DISCONNECTED : ConnectionStatus.USER_DISCONNECTED;
+
+    await this.changeConnectionStatus(newStatus, async () => {
+        if(this.client.isConnected) {
+          await this.client.destroy();
+        }
+    });
+
+    console.log("Disconnected from client");
   }
 
   private startUpdating() {
+    this.idleManager.resetTimer();
     this.updatePresence();
-    this.updateInterval = setInterval(() => this.updatePresence(), 15000);
+    this.updateInterval = setInterval(() => this.updatePresence(), 15333);
   }
 
   private stopUpdating() {
@@ -122,6 +170,7 @@ export default class PresenceManager extends EventEmitter {
 
     // Try re-connecting every 10 seconds.
     this.reconnectionInterval = setInterval(async () => {
+      console.log("Retrying connection...");
       try {
         await this.connectToClient();
         this.startUpdating();
@@ -136,6 +185,7 @@ export default class PresenceManager extends EventEmitter {
 
     // Stop the interval after 2 minutes.
     this.reconnectionTimeout = setTimeout(() => {
+      console.log("Stopping to retry connection");
       clearInterval(this.reconnectionInterval!);
       this.reconnectionInterval = null;
